........
r7243 | jonas | 2007-05-02 16:28:11 +0200 (Wed, 02 May 2007) | 15 lines

  * synchronised with r204 of the universal interfaces (r201-202 were
    tests of the move of the svn server):
...
r204 | jonas | 2007-05-02 16:20:32 +0200 (Wed, 02 May 2007) | 2 lines

  * fixed CFSwapInt64 for FPC

...
r203 | gale | 2007-04-18 15:08:43 +0200 (Wed, 18 Apr 2007) | 5 lines

As a temporary trial fix, fixed record field identifier related
redeclaration warnings by declaring and using additional type declarations
with a "_GAP_Private_field_type_fix" suffix added to conflicting record
field type identifiers.

........

git-svn-id: branches/fixes_2_2@7244 -

+++ packages/extra/univint/CFByteOrders.pas
@@ -190,7 +190,7 @@ end;
 
 function CFSwapInt64( arg: UInt64 ): UInt64; inline;
 begin
-	CFSwapInt64 := (CFSwapInt32( arg and $FFFFFFFF ) shl 32) or CFSwapInt32( (arg shr 32) and $FFFFFFFF );
+	CFSwapInt64 := (UInt64(CFSwapInt32( arg and $FFFFFFFF )) shl 32) or CFSwapInt32( (arg shr 32) and $FFFFFFFF );
 end;
 
 {$ifc TARGET_RT_BIG_ENDIAN}
